[Core] Prevent crashes (#242)
* [Core] Prevent crashes

* [ModSupport] Fixes crash when run.bat does not get created

* [ModManager] Fixes crash when decompressing archive

* [ModSupport] Replaces symlink exception with message box

* [ModSupport] Adds additional try for symbolic links

+++ FrostyModSupport/FrostyModExecutor.cs
@@ -320,6 +320,11 @@ namespace Frosty.ModSupport
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
 
+                                if (!archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
+
                                 archiveData[existingEntry.Sha1].RefCount--;
                                 if (archiveData[existingEntry.Sha1].RefCount == 0)
                                     archiveData.TryRemove(existingEntry.Sha1, out _);
@@ -412,6 +417,11 @@ namespace Frosty.ModSupport
                                     return;
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
+
+                                if (!archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
 
                                 archiveData[existingEntry.Sha1].RefCount--;
                                 if (archiveData[existingEntry.Sha1].RefCount == 0)
@@ -515,6 +525,11 @@ namespace Frosty.ModSupport
                                 ChunkAssetEntry existingEntry = modifiedChunks[guid];
                                 if (existingEntry.Sha1 == resource.Sha1)
                                     return;
+
+                                if (!archiveData.ContainsKey(existingEntry.Sha1))
+                                {
+                                    return;
+                                }
 
                                 archiveData[existingEntry.Sha1].RefCount--;
                                 if (archiveData[existingEntry.Sha1].RefCount == 0)
@@ -743,6 +758,11 @@ namespace Frosty.ModSupport
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
 
+                        if (!archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            continue;
+                        }
+
                         archiveData[existingEntry.Sha1].RefCount--;
                         if (archiveData[existingEntry.Sha1].RefCount == 0)
                             archiveData.TryRemove(existingEntry.Sha1, out _);
@@ -794,6 +814,11 @@ namespace Frosty.ModSupport
                         ResAssetEntry existingEntry = modifiedRes[name];
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
+
+                        if (!archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            continue;
+                        }
 
                         archiveData[existingEntry.Sha1].RefCount--;
                         if (archiveData[existingEntry.Sha1].RefCount == 0)
@@ -848,6 +873,11 @@ namespace Frosty.ModSupport
                         ChunkAssetEntry existingEntry = modifiedChunks[chunkId];
                         if (existingEntry.Sha1 == resource.GetValue<Sha1>("sha1"))
                             continue;
+
+                        if (!archiveData.ContainsKey(existingEntry.Sha1))
+                        {
+                            continue;
+                        }
 
                         archiveData[existingEntry.Sha1].RefCount--;
                         if (archiveData[existingEntry.Sha1].RefCount == 0)
@@ -1245,8 +1275,13 @@ namespace Frosty.ModSupport
                     FrostyMessageBox.Show(reason + "\r\n\r\nShortly you will be prompted for elevated privileges, this is required to create symbolic links between the original data and the new modified data. Please ensure that you accept this to avoid any issues.", "Frosty Toolsuite");
                     if (!RunSymbolicLinkProcess(cmdArgs))
                     {
-                        Directory.Delete(modDataPath, true);
-                        throw new FrostySymLinkException();
+                        FrostyMessageBox.Show("Frosty needs to generate symbolic links, please ensure that you accept this so you don't have to regenerate ModData.", "Frosty Editor");
+                        if (!RunSymbolicLinkProcess(cmdArgs))
+                        {
+                            Directory.Delete(modDataPath, true);
+                            FrostyMessageBox.Show("One ore more symbolic links could not be created, please restart tool as Administrator and ensure your storage drive is formatted to NTFS (not exFAT).", "Frosty Editor");
+                            return -1;
+                        }
                     }
                 }
 
@@ -2144,7 +2179,10 @@ namespace Frosty.ModSupport
             ExecuteProcess("cmd.exe", "/C \"" + AppDomain.CurrentDomain.BaseDirectory + "\\run.bat\"", true, true);
 
             // delete batch
-            File.Delete("run.bat");
+            if (File.Exists("run.bat"))
+            {
+                File.Delete("run.bat");
+            }
 
             // validate
             foreach (SymLinkStruct arg in cmdArgs)
@@ -2181,10 +2219,16 @@ namespace Frosty.ModSupport
                     process.StartInfo.Verb = "runas";
                 }
 
-                process.Start();
+                try
+                {
+                    process.Start();
 
-                if (waitForExit)
-                    process.WaitForExit();
+                    if (waitForExit)
+                        process.WaitForExit();
+                }
+                catch
+                {
+                }
             }
         }
